fix(mcp): run the scenario source-upload guard on the production local-branch collector
collectLocalBranchMetadata only called the local assertSourceUploadDisabled
helper, which checks the LOOPOVER_UPLOAD_SOURCE env flag and nothing else. The
real forbidden-source-upload-key / oversized-content scan lived in
assertScenarioLocalBranchInputSafe (src/scenarios/input-model.ts) but had no
production importer, so source-content fields smuggled into the metadata-only
scenario input were never refused on the actual entry point.

Move that guard into @loopover/engine so both the backend scenario-input model
(re-exported, public surface unchanged) and the loopover-mcp collector share
one scan, wire it into collectLocalBranchMetadata in place of the weaker env
check, and drop the now-dead assertSourceUploadDisabled. Add a regression test
asserting the mcp collector itself rejects forbidden-key, oversized, and
non-array changedFiles payloads.

Closes #8884
